// src/types.ts
export type EventChecklistTask = {
  id: string;
  phase: 'Phase 1' | 'Phase 2' | 'Phase 3' | 'Phase 4' | 'Phase 5 - Post Event' | 'Phase 5 - Promotions';
  taskName: string;
  countdownDays: number;
  deadline: string;
  dateCompleted?: string;
  sampleLinks?: string;
  actualLinks?: string;
  status: 'Please Select' | 'In Progress' | 'Completed' | 'Blocked' | 'Not Applicable';
  notes?: string;
};